pyproject.toml
[project]
name = "web-scraper-voice-agent"
version = "0.1.0"
description = "A web scraper with voice agent capabilities"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.1",
    "beautifulsoup4>=4.12.2",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [] 